pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sclsd"
dynamic = ["version"]
description = "Latent State Dynamics (LSD) for single-cell trajectory inference via neural ODE gradient flow"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "LSD Development Team" }
]
keywords = [
    "single-cell",
    "trajectory-inference",
    "cell-differentiation",
    "deep-learning",
    "neural-ode",
    "pyro",
    "variational-inference",
    "Waddington-landscape",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "pyro-ppl>=1.8.0",
    "torchdiffeq>=0.2.0",
    "scanpy>=1.9.0",
    "anndata>=0.9.0",
    "cellrank>=2.0.0",
    "numpy>=1.23.0,<2.0.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.60.0",
    "leidenalg",
    "umap-learn>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "furo>=2024.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "sphinx-copybutton>=0.5.0",
    "nbsphinx>=0.9.0",
    "myst-parser>=2.0.0",
]
all = ["sclsd[dev,docs]"]

[project.urls]
Homepage = "https://github.com/csglab/sclsd"
Documentation = "https://csglab.github.io/sclsd"
Repository = "https://github.com/csglab/sclsd"
Issues = "https://github.com/csglab/sclsd/issues"

[tool.hatch.version]
path = "src/sclsd/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/sclsd"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/notebooks",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["sclsd"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
    "integration: integration tests",
    "reproducibility: reproducibility tests",
    "parity: comparison with original LSD implementation",
]
